Fix 1.4 support, events not working
